update4

+++ src/main/java/org/springMVChibernateCRUD/dao/UserDAO.java
@@ -8,12 +8,12 @@ public interface UserDAO {
 
     List<User> getListUsers();
 
-    User getUser(int id);
+    User getUser(Long id);
 
     void addUser(User user);
 
-    void update(int id, User user);
+    void update(Long id, User user);
 
-    void delete(int id);
+    void delete(Long id);
 
 }
